// MiLicenciaCentro/src/app/interfaces/compraPin/PagoPin.ts
import { PasosCompraPin } from 'src/app/enums/forms/PasosCompraPin';
import { MedioPago } from "src/app/enums/PinesOlimpia/TipoPago";
import { CostoCuota } from "src/app/interfaces/compraPin/CostoCuota";
import { Categoria } from "src/app/interfaces/cotizacion/Categoria";
import { Centro } from "src/app/interfaces/cotizacion/Centro";
import { Convenio } from "src/app/interfaces/cotizacion/Convenio";
import { DiscriminadoValorPin } from "src/app/interfaces/cotizacion/DiscriminadoValorPin";
import { Analytics } from 'src/app/interfaces/Otros/Analytics';
import { Banco } from "src/app/interfaces/pago/SecurePay/Banco";
import { PilotoCea } from 'src/app/interfaces/pago/SecurePay/PilotoCea';
import { Bancos } from "src/app/interfaces/pago/TuPago/Banco";
import { TipoDocumentoPtesaDTO } from '../cotizacion/TipoDocumentoPtesaDTO';
import { DatosBasicos } from "./DatosBasicos";
import { DatosFacturaElectronica } from './DatosFacturaElectronica';
import { tipoDocumento } from '../pago/FacturacionElectronica/TipoDocumentoResponse';
import { tipoPersona } from '../pago/FacturacionElectronica/TipoPersonaResponse';

export interface PagoPin{
    usuario: DatosBasicos;
    comprador: DatosBasicos;
    compradorFE:DatosFacturaElectronica;
    clienteCompra: number;//cea o crc
    tipoTramite: number | null;
    tipoTramite2: number | null;
    opcionTramite: number | null;
    validacionMapa:boolean;
    centroSeleccionado?: Centro | null;
    conveniosPagoEfectivo: Convenio[];
    conveniosMediosPago: Convenio[];
    mediosPago: MedioPago[];
    mediosPagoTodos: MedioPago[];
    bancosPSE: Bancos[];
    bancosPSESecure: Banco[];
    costoCuotas: CostoCuota[];
    costoCuotasSeleccionadas: number;
    valorDiscriminadoCotizacion: DiscriminadoValorPin | null;
    valorDiscriminadoCotizacionAprox: DiscriminadoValorPin | null;
    categoria: string;
    categoria1: string;
    categoria2: string;
    categoriasActual: string | null ;
    edadAspirante: number;
    tipoRecaudoCtrl:number | null;
    tipoPagoEfectivo: number | null;
    nomenclaturaIdentificacion: string;
    tipoPersonaPSE: string | null;
    tipoBancoPSE: string | null;
    cuotas: number | null;
    referenciaGenerada: string;
    pasarelaTuPago: boolean;
    urlCentro: boolean;
    categorias?: Categoria[];
    obtenerPagoCrc: boolean;
    pasoCotizacion:PasosCompraPin;
    tiposDeDocumento: TipoDocumentoPtesaDTO[];
    tiposDeDocumentoFE:tipoDocumento[];
    tipoPersonaFE:tipoPersona[];
    host: string;
    validarlogueado: boolean;
    tieneConvenios: boolean;
    centroIpValidacion: boolean;
    centroIpVolver: boolean;
    categoriasCentroIp: Categoria[] | undefined;
    plantillas: Map<String, number> | undefined;
    analytics: Analytics;
    pilotoCEA: PilotoCea | undefined;
    registroMiPerfil: number;
    categoriasCrc: Categoria[];
    categoriasCea: Categoria[];
    tramiteInstructor:boolean;
    checkDaviplata:boolean;
    checkFacturaElectronica:boolean;
    checkFacturaElectronicaResumen:boolean;
}